+++ src/main/java/com/example/ecommerce/repository/CartRepository.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 
 @Repository
 public interface CartRepository extends JpaRepository<Cart, Integer> {
-    Cart findByUserId(Integer userId);
+    Cart findByUserId(Long userId);
     Cart save(Cart cart);
     void deleteById(Integer id);
